backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "retail-erp-backend"
version = "1.0.0"
description = "Retail ERP Platform Backend API"
authors = [
    {name = "Development Team", email = "dev@company.com"}
]
license = {text = "Private"}
requires-python = ">=3.9"
dependencies = [
    "Django>=5.0",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.0.0",
    "python-dotenv>=1.0.0",
    "psycopg[binary]>=3.1.0",
    "pillow>=10.0.0",
    "django-filter>=23.0",
    "drf-spectacular>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-django>=4.5.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88